tlshd: Combine PSK and x.509 support into a single file
Clean up: Reorganize some TLS-specific code into one source file.
