+++ src/tlshd/Makefile.am
@@ -26,7 +26,7 @@ sbin_PROGRAMS		= tlshd
 tlshd_CFLAGS		= -Werror -Wall -Wextra $(LIBGNUTLS_CFLAGS) \
 			   $(LIBKEYUTILS_CFLAGS) $(GLIB_CFLAGS)
 tlshd_SOURCES		= config.c handshake.c keyring.c ktls.c log.c main.c \
-			  psk.c tlshd.h x509.c
+			  tlshd.h x509.c
 tlshd_LDADD		= $(LIBGNUTLS_LIBS) $(LIBKEYUTILS_LIBS) $(GLIB_LIBS) \
 			  $(LIBBSD_LIBS)
