+++ src/tlshd/tlshd.h
@@ -67,12 +67,10 @@ extern void tlshd_gnutls_audit_func(gnutls_session_t session, const char *msg);
 
 void tlshd_log_gerror(const char *msg, GError *error);
 
-/* psk.c */
-extern void tlshd_client_psk_handshake(int sock, const char *tlshd_peername);
-
 /* x509.c */
 extern void tlshd_client_anon_handshake(int sock, const char *peername);
 extern void tlshd_client_x509_handshake(int sock, const char *peername);
+extern void tlshd_client_psk_handshake(int sock, const char *tlshd_peername);
 
 #if !defined(AF_TLSH)
